module structure improvements
