+++ src/background/icon/iconSetter.js
