+++ src/manifest.json
@@ -1,7 +1,7 @@
 {
   "manifest_version": 3,
   "name": "Save my Phind - AI Conversation Exporter",
-  "version": "1.7.5",
+  "version": "1.7.6",
   "description": "🚀 Export your Phind.com search threads into markdown files!\n",
   "icons": {
     "16": "img/icons/icon-16.png",
@@ -10,7 +10,7 @@
   },
   "permissions": [
     "scripting",
-    "tabs"
+    "activeTab"
   ],
   "host_permissions": [
     "https://*.phind.com/*"
